+++ xelis_common/src/utils.rs
@@ -30,8 +30,8 @@ pub fn get_current_time_in_seconds() -> u64 {
 }
 
 // return timestamp in milliseconds
-pub fn get_current_time_in_millis() -> u128 {
-    get_current_time().as_millis()
+pub fn get_current_time_in_millis() -> u64 {
+    get_current_time().as_millis() as u64
 }
 
 // Format any coin value using the requested decimals count
